9/n: move common functions to another file
